Moved plot_qscore_histogram logic to cpp (cherry picked from commit ef2e8eb)
