+++ src/interop/CMakeLists.txt
@@ -17,7 +17,8 @@ set(SRCS
         logic/plot/plot_by_lane.cpp
         logic/plot/plot_flowcell_map.cpp
         logic/plot/plot_qscore_heatmap.cpp
-        logic/plot/plot_sample_qc.cpp)
+        logic/plot/plot_sample_qc.cpp
+        logic/plot/plot_qscore_histogram.cpp)
 
 set(HEADERS
         ../../interop/model/summary/cycle_state_summary.h
